Removed reflection and externalized the API for Botania
