+++ src/main/java/theflogat/technomancy/common/tiles/thaumcraft/machine/TileManaFabricator.java
@@ -7,7 +7,7 @@ import net.minecraftforge.common.util.ForgeDirection;
 import theflogat.technomancy.common.blocks.base.TMBlocks;
 import theflogat.technomancy.common.tiles.base.TileMachineBase;
 import theflogat.technomancy.lib.Rate;
-import theflogat.technomancy.lib.compat.Botania;
+import vazkii.botania.api.BotaniaAPI;
 import vazkii.botania.api.mana.IManaPool;
 
 public class TileManaFabricator extends TileMachineBase implements IManaPool {
@@ -68,11 +68,7 @@ public class TileManaFabricator extends TileMachineBase implements IManaPool {
 	
 	public void renderHUD(Minecraft mc, ScaledResolution res) {
 		int color = 0x660000FF;
-		try {
-			Botania.drawHUD.invoke(null, color, mana, maxMana, TMBlocks.manaFabricator.getLocalizedName(), res);
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
+		BotaniaAPI.internalHandler.drawSimpleManaHUD(color, mana, maxMana, TMBlocks.manaFabricator.getLocalizedName(), res);
 	}
 	
 	@Override
